refactor(s3): move s3 service to global package
